WIP: Login in price calculator
