+++ apps/store/src/components/CartPage/CartPage.tsx
@@ -2,19 +2,14 @@ import { datadogLogs } from '@datadog/browser-logs'
 import styled from '@emotion/styled'
 import { useTranslation } from 'next-i18next'
 import { useRouter } from 'next/router'
-import { FormEventHandler, ReactNode, useEffect, useState } from 'react'
-import { Button, Heading, mq, Space, Text, theme } from 'ui'
+import { ReactNode, useEffect } from 'react'
+import { Heading, mq, Space, Text, theme } from 'ui'
+import { BankIdLogin } from '@/components/BankIdLogin'
 import { CampaignCodeList } from '@/components/CartInventory/CampaignCodeList'
 import { CartEntryItem } from '@/components/CartInventory/CartEntryItem'
 import { CartEntryList } from '@/components/CartInventory/CartEntryList'
 import { CostSummary } from '@/components/CartInventory/CostSummary'
-import {
-  ShopSessionAuthenticationStatus,
-  useShopSessionAuthenticateMutation,
-} from '@/services/apollo/generated'
-import { loginMemberSeBankId } from '@/services/authApi/login'
-import { exchangeAuthorizationCode } from '@/services/authApi/oauth'
-import { saveAccessToken } from '@/services/authApi/persist'
+import { ShopSessionAuthenticationStatus } from '@/services/apollo/generated'
 import { ShopSession } from '@/services/shopSession/ShopSession.types'
 import { useShopSession } from '@/services/shopSession/ShopSessionContext'
 import { useTracking } from '@/services/Tracking/useTracking'
@@ -143,11 +138,16 @@ const CartNextStep = ({ shopSession }: { shopSession: ShopSession }) => {
     ShopSessionAuthenticationStatus.AuthenticationRequired
   ) {
     return (
-      <BankIdLogin
-        shopSessionId={shopSession.id}
-        ssn={shopSession.customer.ssn}
-        onComplete={handleAuthSuccess}
-      />
+      <>
+        <Text>
+          Looks like you&quot;re returning member. You need to login before proceeding to checkout
+        </Text>
+        <BankIdLogin
+          shopSessionId={shopSession.id}
+          ssn={shopSession.customer.ssn}
+          onCompleted={handleAuthSuccess}
+        />
+      </>
     )
   } else {
     return (
@@ -156,42 +156,6 @@ const CartNextStep = ({ shopSession }: { shopSession: ShopSession }) => {
       </ButtonNextLink>
     )
   }
-}
-
-// TODO: Make it reusable between cart and price calculator prompt
-// TODO: Localize texts
-const BankIdLogin = (props: { shopSessionId: string; ssn: string; onComplete: () => void }) => {
-  const [state, setState] = useState<'IDLE' | 'PROGRESS' | 'ERROR'>('IDLE')
-  const { shopSessionId } = props
-  const [authenticateShopSession] = useShopSessionAuthenticateMutation({
-    variables: { shopSessionId },
-  })
-  const handleSubmit: FormEventHandler = async (event) => {
-    event.preventDefault()
-    setState('PROGRESS')
-    try {
-      const authorizationCode = await loginMemberSeBankId(props.ssn)
-      const accessToken = await exchangeAuthorizationCode(authorizationCode)
-      saveAccessToken(accessToken)
-      await authenticateShopSession()
-      props.onComplete()
-    } catch (error) {
-      datadogLogs.logger.warn('Failed to authenticate', { error })
-      setState('ERROR')
-    }
-  }
-  return (
-    <form onSubmit={handleSubmit}>
-      <Text>
-        Looks like you&quot;re returning member. You need to login before proceeding to checkout
-      </Text>
-      <Button type="submit" loading={state !== 'IDLE'}>
-        BankID Login
-      </Button>
-      {state === 'PROGRESS' && <Text>Pleas open BankID app now</Text>}
-      {state === 'ERROR' && <Text>Something went wrong</Text>}
-    </form>
-  )
 }
 
 const HorizontalLine = styled.hr({
